% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accGPS.R
\name{accGPS}
\alias{accGPS}
\title{Calculate acceleration between GPS points}
\usage{
accGPS(timeVec, speedVec)
}
\arguments{
\item{timeVec}{Time vector of the GPS data points in \dQuote{Y-m-dTH:M:SZ} format.}

\item{speedVec}{Vector of velocities (in m/s) between GPS data points (output of the \code{\link{speedGPS}}).}
}
\value{
Function returns the value or vector of calculated acceleration in m/s^2.
}
\description{
This function calculates the acceleration in m/s^2 between pair or vector of GPS points based on the time and velocity between them.
}
\examples{
data_file <- system.file("extdata", "myGPSData.csv", package="analyzeGPS")
gps <- readGPS(data_file)
d <- distanceGPS(lat1 = gps$lat[1:(length(gps$lat)-1)], lon1 = gps$lon[1:(length(gps$lon)-1)],
lat2 = gps$lat[2:length(gps$lat)], lon2 = gps$lon[2:length(gps$lon)])
speed <- speedGPS(gps$time, d)
acc <- accGPS(gps$time, speed)
}
\keyword{GPS}
\keyword{acceleration}
\keyword{calculation}
\keyword{data,}

