% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceGPS.R
\name{distanceGPS}
\alias{distanceGPS}
\title{Calculate distance between GPS points}
\usage{
distanceGPS(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{latitude data of the first point (or vector of points).}

\item{lon1}{longitude data of the first point (or vector of points).}

\item{lat2}{latitude data of the second point (or vector of points).}

\item{lon2}{longitude data of the second point (or vector of points).}
}
\value{
Function returns the value or vector of calculated distances in meters.
}
\description{
This function calculates the distance in meters between pair or vector of GPS points described with latitude and longitude. The calculation considers the Earth radius by applying the Haversine formula. The number of elements of the output is one less than the number of input (lon, lat).
}
\examples{
data_file <- system.file("extdata", "myGPSData.csv", package="analyzeGPS")
gps <- readGPS(data_file)
d <- distanceGPS(lat1 = gps$lat[1:(length(gps$lat)-1)], lon1 = gps$lon[1:(length(gps$lon)-1)],
lat2 = gps$lat[2:length(gps$lat)], lon2 = gps$lon[2:length(gps$lon)])
}
\keyword{GPS}
\keyword{calculation}
\keyword{data,}
\keyword{distance}

