% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradeGPS.R
\name{gradeGPS}
\alias{gradeGPS}
\title{Calculate grade between GPS points}
\usage{
gradeGPS(eleVec, distanceVec)
}
\arguments{
\item{eleVec}{Vector of elevations of GPS data points.}

\item{distanceVec}{Vector of distances between GPS data points (output of the \code{distanceGPS}, padded with an extra zero at the beginning (see vignette of this package)).}
}
\value{
Function returns the value or vector of calculated grade as the ratio between the rise and distance traveled.
}
\description{
This function calculates the grade between pair or vector of GPS points as the ratio between the rise and distance traveled.
}
\examples{
data_file <- system.file("extdata", "myGPSData.csv", package="analyzeGPS")
gps <- readGPS(data_file)
d <- distanceGPS(lat1 = gps$lat[1:(length(gps$lat)-1)], lon1 = gps$lon[1:(length(gps$lon)-1)],
lat2 = gps$lat[2:length(gps$lat)], lon2 = gps$lon[2:length(gps$lon)])
grade <- gradeGPS(gps$ele, d)
}
\keyword{GPS}
\keyword{calculation}
\keyword{data,}
\keyword{grade}

