% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeGPS.R
\docType{data}
\name{myGPSData}
\alias{myGPSData}
\title{myGPSData}
\format{A data frame with 7771 rows and 5 variables}
\description{
A dataset containing some sample GPS data.
The variables are as follows:
}
\details{
\itemize{
 \item lon: longitude data
 \item lat: latitude data
 \item ele: elevation data
 \item time: GPS time stamp - GMT time zone
 \item tz_CEST: time stamp converted to CEST time zone
}
}

