% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGPS.R
\name{readGPS}
\alias{readGPS}
\title{Read GPS data file}
\usage{
readGPS(NAME = "filename")
}
\arguments{
\item{NAME}{The input parameter is the string path to the selected GPS file (csv format).}
}
\value{
Function returns the gps data frame.
}
\description{
This function reads the acquired GPS data in csv format and sorts it into a data frame.
}
\examples{
data_file <- system.file("extdata", "myGPSData.csv", package="analyzeGPS")
gps <- readGPS(data_file)
}
\keyword{GPS}
\keyword{data,}
\keyword{read}

