% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/speedGPS.R
\name{speedGPS}
\alias{speedGPS}
\title{Calculate velocity/speed between GPS points}
\usage{
speedGPS(timeVec, distanceVec)
}
\arguments{
\item{timeVec}{Time vector of the GPS data points in \dQuote{Y-m-dTH:M:SZ} format.}

\item{distanceVec}{Vector of distances between GPS data points (output of the \code{\link{distanceGPS}}).}
}
\value{
Function returns the value or vector of calculated speed in m/s.
}
\description{
This function calculates the speed in meters per second between pair or vector of GPS points based on the time and distance between them.
}
\examples{
data_file <- system.file("extdata", "myGPSData.csv", package="analyzeGPS")
gps <- readGPS(data_file)
d <- distanceGPS(lat1 = gps$lat[1:(length(gps$lat)-1)], lon1 = gps$lon[1:(length(gps$lon)-1)],
lat2 = gps$lat[2:length(gps$lat)], lon2 = gps$lon[2:length(gps$lon)])
speed <- speedGPS(gps$time, d)
}
\keyword{GPS}
\keyword{calculation}
\keyword{data,}
\keyword{velocity}

